backend/requirements.txt
# Web 框架
fastapi==0.109.2
uvicorn[standard]==0.27.1

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# pgvector
pgvector==0.2.5

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# RAG 与 LLM
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24
openai==1.12.0
tiktoken==0.6.0

# 文档解析
pypdf==4.0.1
python-docx==1.1.0
# 工具
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
